requirements.txt
mysql-connector-python==8.2.0
bcrypt==4.1.2
pillow==9.5.0
tkcalendar

# jalankan ini yaaa
# pip install -r requirements.txt